+++ lib/flutter_ddi.dart
@@ -4,11 +4,11 @@ export 'package:dart_ddi/dart_ddi.dart';
 
 export 'src/extensions/flutter_ddi_context.dart';
 export 'src/extensions/flutter_ddi_router_extension.dart';
+export 'src/features/interceptors/flutter_ddi_interceptor.dart';
+export 'src/features/interceptors/interceptor_result.dart';
+export 'src/features/interceptors/module_interceptor.dart';
 export 'src/features/listener/flutter_ddi_event_listener.dart';
 export 'src/features/listener/flutter_ddi_stream_listener.dart';
-export 'src/features/middleware/flutter_ddi_middleware.dart';
-export 'src/features/middleware/middleware.dart';
-export 'src/features/middleware/middleware_result.dart';
 export 'src/features/state/event_listener_state.dart';
 export 'src/features/state/stream_listener_state.dart';
 export 'src/interfaces/error_module_interface.dart';