+++ lib/src/features/interceptors/module_interceptor.dart
@@ -3,14 +3,14 @@ import 'dart:async';
 import 'package:flutter/foundation.dart';
 import 'package:flutter_ddi/flutter_ddi.dart';
 
-final class Middleware<MiddlewareT extends FlutterDDIMiddleware> {
-  /// Use to register a new Middleware.
-  factory Middleware.of({
-    required ScopeFactory<MiddlewareT> factory,
+final class ModuleInterceptor<InterceptorT extends FlutterDDIInterceptor> {
+  /// Use to register a new Interceptor.
+  factory ModuleInterceptor.of({
+    required ScopeFactory<InterceptorT> factory,
     Object? qualifier,
     FutureOr<bool> Function()? canRegister,
   }) {
-    return Middleware._(
+    return ModuleInterceptor._(
       qualifier: qualifier ?? factory.type,
       factory: factory,
       canRegister: canRegister,
@@ -18,34 +18,34 @@ final class Middleware<MiddlewareT extends FlutterDDIMiddleware> {
     );
   }
 
-  /// Use to reuse a Middleware already registered.
-  factory Middleware.from({Object? qualifier}) {
-    return Middleware._(
-      qualifier: qualifier ?? MiddlewareT,
+  /// Use to reuse a Interceptor already registered.
+  factory ModuleInterceptor.from({Object? qualifier}) {
+    return ModuleInterceptor._(
+      qualifier: qualifier ?? InterceptorT,
       isFactory: false,
     );
   }
 
-  Middleware._({
+  ModuleInterceptor._({
     required this.qualifier,
     required this.isFactory,
     this.factory,
     this.canRegister,
   });
 
-  /// The type of the Middleware.
+  /// The type of the Interceptor.
   final Object qualifier;
-  final ScopeFactory<MiddlewareT>? factory;
+  final ScopeFactory<InterceptorT>? factory;
   final FutureOr<bool> Function()? canRegister;
   final bool isFactory;
 
-  /// Registers the middleware asynchronously.
+  /// Registers the interceptor asynchronously.
   ///
-  /// If the middleware is a factory, it will be registered with the provided
+  /// If the interceptor is a factory, it will be registered with the provided
   /// factory method, qualifier, and conditional registration function.
   ///
-  /// Returns the qualifier of the registered middleware if provided, otherwise
-  /// returns the type of the middleware.
+  /// Returns the qualifier of the registered interceptor if provided, otherwise
+  /// returns the type of the interceptor.
   @nonVirtual
   Future<Object> register() async {
     if (isFactory) {
@@ -59,10 +59,10 @@ final class Middleware<MiddlewareT extends FlutterDDIMiddleware> {
     return qualifier;
   }
 
-  /// Destroys the middleware from the container.
+  /// Destroys the interceptor from the container.
   ///
-  /// This method should be used when you want to remove the middleware from the container.
-  /// This is useful when you want to remove a middleware from the container after it has been used.
+  /// This method should be used when you want to remove the interceptor from the container.
+  /// This is useful when you want to remove a interceptor from the container after it has been used.
   /// If [isFactory] is false, this method does nothing.
   @nonVirtual
   Future<void> destroy() async {